Add Mastercard
The initial call to action in the country selector on the US site calls to select your location. But they also have a subtext for missing countries and regions with "Don't see a country or region listed? Visit our global site."

When you visit the german site the initial call to action is to select your country.

What do you think how should we classify Mastercard?